+++ oap-server/server-query-plugin/query-graphql-plugin/src/main/java/org/apache/skywalking/oap/query/graphql/GraphQLQueryConfig.java
@@ -30,6 +30,5 @@ import org.apache.skywalking.oap.server.library.module.ModuleConfig;
 public class GraphQLQueryConfig extends ModuleConfig {
     private boolean enableLogTestTool;
     private int maxQueryComplexity = 1000;
-    private boolean enableUpdateUITemplate = false;
     private boolean enableOnDemandPodLog = false;
 }